requirements.txt
langchain
langchain_openai
langchain_core
python-dotenv
streamlit
langchain_community
langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf
faiss-cpu
pandas
langchain_ollama
arxiv
wikipedia
langchainhub
watchdog 
PyPDF2 
pandas
langchain-google-genai
python-jose
requests
pymongo
huggingface_hub
google-genai
# google-genai Pillow
langchain-huggingface
huggingface_hub
transformers
accelerate
bitsandbytes
google-cloud-speech
google-generativeai
react-loading-skeleton